change int -> size_t for tapestats in all examples

+++ ADOL-C/examples/additional_examples/lufact/LUsolve.cpp
@@ -123,7 +123,7 @@ int main() { /*-----------------------------------------------------------------
 
     /*------------------------------------------------------------------------*/
     /* Tape statistics */
-    int tape_stats[STAT_SIZE];
+    size_t tape_stats[STAT_SIZE];
     tapestats(tag,tape_stats);
 
     fprintf(stdout,"\n    independents            %d\n",tape_stats[NUM_INDEPENDENTS]);
